move app related components into component/app folder
