+++ packages/editor-web-ui/src/app/page.tsx
@@ -1,4 +1,4 @@
-import { App } from '../components/app';
+import { App } from '../components/app/app';
 
 export default function HomePage() {
   return <App />;
