+++ packages/editor-web-ui/src/components/app/app-sidebar-project-switcher.tsx
